# .well-known/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Text to Speech API",
    "version": "1.0.0",
    "description": "An API to convert text to speech using Eleven Labs API and providing a download link for the generated audio file."
  },
  "servers": [
    {
      "url": "http://116.203.93.2:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/generate-audio": {
      "post": {
        "operationId": "GenerateAudio",
        "summary": "Generate an audio file from text",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "The text to convert to speech"
                  }
                },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Audio file generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "downloadLink": {
                      "type": "string",
                      "description": "Link to download the generated audio file"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error occurred while generating audio file"
          }
        }
      }
    },
    "/download/{fileName}": {
      "get": {
        "operationId": "DownloadAudio",
        "summary": "Download the generated audio file",
        "parameters": [
          {
            "in": "path",
            "name": "fileName",
            "required": true,
            "description": "The name of the file to download",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File downloaded successfully"
          },
          "500": {
            "description": "Error occurred while downloading the file"
          }
        }
      }
    }
  }
}